[FIX] mix_precision

+++ utils.py
@@ -60,7 +60,8 @@ def get_optimizer(args):
 def gram_matrix(input_tensor):
     result = tf.linalg.einsum('bijc,bijd->bcd', input_tensor, input_tensor)
     input_shape = tf.shape(input_tensor)
-    num_locations = tf.cast(input_shape[1]*input_shape[2], tf.float32)
+    # num_locations = tf.cast(input_shape[1]*input_shape[2], tf.float32)
+    num_locations = input_shape[1]*input_shape[2]
     return result/(num_locations)
